back/requirements.txt
scrapy
celery
djangorestframework
django
celery
redis
scrapyd
scrapyd-client
psycopg2